agreement: Fix cleanup of ABBA instance
Replying with FINISH(v) messages to any message directed at the instance
does not affect correctness but leads to infinite chatter between nodes,
for all agreement rounds.

This commit introduces a new special agreement message
FinishAbbaInstance, that is meant specifically for replicas to help
others catch up: it is forwarded by the agreement module as a FINISH(v)
messsage to the underlying instance, but, crucially, it does not trigger
a reply from up-to-date replicas.