reliablenet: Give up on concurrency
MarkModuleMsgsRecvd races with SendMessage leaving us with pending
messages that are never acknowledged.

We could introduce explicit queue creation and destruction, but this
would complicate the usage of reliablenet. Another option would be to
propagate metadata (such as message IDs) to allow blind acknowledgements.

The heavy lifting is performed by the underlying net implementation, so
this module should be enough as is.