alea: Really limit agreement delay to MaxAgreementDelay
Previously we just scheduled a timer to ensure the director ran its
conditions with the period of MaxAgreementDelay.

The problem is if the broadcast estimate is temporarily inflated.
Agreement will hang, waiting for the large estimated period of time to
pass, and no new broadcasts are happening so this estimate does not
lower for some time.