alea/director,agreement: dynamically adjust bc estimate margin
The bc estimate margin is now doubled whenever a batch from the current
node is rejected in agreement (to give more time for the broadcast to be
completed in remote replicas).

To tighten the margin (which yields better performance) this margin is
lowered to approach the time between agreement input and a quorum of 1s
in agreement.